desafio2-volumes-persistencia/client/requirements.txt
psycopg2-binary==2.9.9
colorama==0.4.6
